move customMeasurementOverlay to a separate file
